+++ packages/app/src/pages/actueel/gemeente/[code].tsx
@@ -34,7 +34,7 @@ import { selectVaccineCoverageData } from '~/domain/vaccine/data-selection/selec
 import { useAgegroupLabels } from '~/domain/vaccine/logic/use-agegroup-labels';
 import { useIntl } from '~/intl';
 import { useFeature } from '~/lib/features';
-import { getWarning } from '~/queries/create-elements-query';
+import { getWarning } from '~/queries/get-elements-query';
 import { getTopicalPageData } from '~/queries/get-topical-page-data';
 import {
   createGetStaticProps,
