+++ packages/app/src/pages/gemeente/[code]/vaccinaties.tsx
@@ -27,7 +27,7 @@ import {
   getArticleParts,
   getLinkParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
