+++ packages/app/src/pages/index.tsx
@@ -38,7 +38,7 @@ import { TopicalSectionHeader } from '~/domain/topical/topical-section-header';
 import { selectVaccineCoverageData } from '~/domain/vaccine/data-selection/select-vaccine-coverage-data';
 import { useIntl } from '~/intl';
 import { useFeature } from '~/lib/features';
-import { getWarning } from '~/queries/create-elements-query';
+import { getWarning } from '~/queries/get-elements-query';
 import { getTopicalPageData } from '~/queries/get-topical-page-data';
 import {
   createGetStaticProps,
