+++ packages/app/src/pages/internationaal/positief-geteste-mensen.tsx
@@ -34,7 +34,7 @@ import {
   getArticleParts,
   getLinkParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
