+++ packages/app/src/pages/internationaal/varianten.tsx
@@ -25,7 +25,7 @@ import {
   getArticleParts,
   getLinkParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
