+++ packages/app/src/pages/landelijk/brononderzoek.tsx
@@ -11,7 +11,7 @@ import { withFeatureNotFoundPage } from '~/lib/features';
 import {
   getArticleParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
