+++ packages/app/src/pages/veiligheidsregio/[code]/positief-geteste-mensen.tsx
@@ -22,14 +22,14 @@ import { VrLayout } from '~/domain/layout/vr-layout';
 import { GNumberBarChartTile } from '~/domain/tested/g-number-bar-chart-tile';
 import { useIntl } from '~/intl';
 import {
-  createElementsQuery,
   ElementsQueryResult,
+  getElementsQuery,
   getTimelineEvents,
-} from '~/queries/create-elements-query';
+} from '~/queries/get-elements-query';
 import {
   getArticleParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
@@ -69,7 +69,7 @@ export const getStaticProps = createGetStaticProps(
       const { locale } = context;
       return `{
        "parts": ${getPagePartsQuery('positiveTestsPage')},
-       "elements": ${createElementsQuery(
+       "elements": ${getElementsQuery(
          'vr',
          ['tested_overall', 'tested_ggd'],
          locale
