+++ packages/app/src/pages/veiligheidsregio/[code]/thuiswonende-ouderen.tsx
@@ -16,7 +16,7 @@ import { useIntl } from '~/intl';
 import {
   getArticleParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
