+++ packages/app/src/pages/veiligheidsregio/[code]/ziekenhuis-opnames.tsx
@@ -21,15 +21,15 @@ import { Layout } from '~/domain/layout/layout';
 import { VrLayout } from '~/domain/layout/vr-layout';
 import { useIntl } from '~/intl';
 import {
-  createElementsQuery,
   ElementsQueryResult,
+  getElementsQuery,
   getTimelineEvents,
-} from '~/queries/create-elements-query';
+} from '~/queries/get-elements-query';
 import {
   getArticleParts,
   getLinkParts,
   getPagePartsQuery,
-} from '~/queries/get-page-parts.query';
+} from '~/queries/get-page-parts-query';
 import {
   createGetStaticProps,
   StaticProps,
@@ -61,11 +61,7 @@ export const getStaticProps = createGetStaticProps(
     }>((context) => {
       return `{
         "parts": ${getPagePartsQuery('hospitalPage')},
-        "elements": ${createElementsQuery(
-          'vr',
-          ['hospital_nice'],
-          context.locale
-        )}
+        "elements": ${getElementsQuery('vr', ['hospital_nice'], context.locale)}
       }`;
     })(context);
